Clickable Game (READ DESCRIPTION)
Here's the deal.  I got clicking stuff to work.  The three files I worked on are ClickableGame.java, SampleBoard.java, and Slot.java.  Slot also needs Card.java.  A board gets drawn which has 4 rows of 10 positions where cards could be placed during the game.  The top row is row 0 and the top 2 rows belong to player 2.  The bottom row is 3, and the bottom 2 rows belong to player 1.  The leftmost card in a row is column 0, the rightmost 9.  Each card position can be clicked on, and the program will access a 2D array of Slot objects that have Card objects stored in them.  Depending on where the Card objects are (hand or field) the program will do something different with them.  If they are in a hand, the program will find the first open slot in that player's field and move the card there.  If it's in the field, combat will happen, but we're not worried about that yet.  I need you guys to make the cards go into the slots, have a card image get drawn whenever it's moved to a slot, and also maybe add a draw function if you want that uses my function for finding the first opening in an array to put a new card into the first open spot of a player's hand.