Combat Works!
There is now an additional column on the right which contains (from top to bottom):
Player 1 HP, A useless button where we can display "Player 1/2 Wins!" at the end of the game, An End Turn Button (not yet functional), and Player 2 HP

If a minion attacks an empty space, it will instead attack your opponent directly!!  So combat works!

What we need now is a turn system (just make a turn variable, and in that bottom most function, so add a condition so top rows only trigger when turn is 1, and bottom rows only trigger when turn is 2.  Make the end turn button switch turn and draw a card.  We also need to find the leftmost open spot in a row so we can add cards to the hand and play minions in the leftmost open spot, instead of directly ahead.  Make sure that when checking for leftmost open spot, you ignore the rightmost column.  We also need mana, and a way to make cards only be abgle to attack once per turn (probably just add a boolean variable to the card object that gets refereshed by the end turn button, as well as a couple of conditions in combat).