Big Functionality Update
I kinda just binge programmed and made functional turns and mana.  There's an image in the assets folder on drive called "EndTurn.png" that you need.  So basically all we still need is drawing a card at the end of the turn, the game stopping when a player dies, a completed card database, and I think that's it??  I'm probably forgetting something.